Updated file structure to move as much as possible to plugins/dev
Updated makefile to build development plugins

Added plugin to hook Tween updates to render loop and added a few camera animations based on it. These don't test if Tween is hooked up but they should as they are a dependency, perhaps create a tween animation plugin that extends camera control?